# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bacchus Marsh Taxi Cab API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CreateBookingBodyVehicleType =
  (typeof CreateBookingBodyVehicleType)[keyof typeof CreateBookingBodyVehicleType];

export const CreateBookingBodyVehicleType = {
  sedan: "sedan",
  suv: "suv",
  silver_service: "silver_service",
  six_seater: "six_seater",
  maxi_taxi: "maxi_taxi",
} as const;

export interface CreateBookingBody {
  name: string;
  phone: string;
  email: string;
  pickupAddress: string;
  dropoffAddress: string;
  vehicleType: CreateBookingBodyVehicleType;
  /**
   * @minimum 1
   * @maximum 13
   */
  passengers: number;
  pickupDate: string;
  pickupTime: string;
  isReturn: boolean;
  returnDate?: string | null;
  returnTime?: string | null;
  estimatedFare?: number | null;
  notes?: string | null;
  distanceKm?: number | null;
  paymentMethod?: "cash" | "card" | "cabcharge" | null;
}

export interface BookingResponse {
  success: boolean;
  message: string;
  bookingId: string;
}

export type FareEstimateBodyVehicleType =
  (typeof FareEstimateBodyVehicleType)[keyof typeof FareEstimateBodyVehicleType];

export const FareEstimateBodyVehicleType = {
  sedan: "sedan",
  suv: "suv",
  silver_service: "silver_service",
  six_seater: "six_seater",
  maxi_taxi: "maxi_taxi",
} as const;

export interface FareEstimateBody {
  distanceKm: number;
  vehicleType: FareEstimateBodyVehicleType;
  pickupDate?: string | null;
  pickupTime?: string | null;
  /** Toll roads detected on the route (e.g. "CityLink", "EastLink", "Westgate Tunnel") */
  tollRoads?: string[] | null;
  /** Optional live traffic multiplier from Google Maps (e.g. 1.25 means traffic is 25% slower than baseline) */
  trafficRatio?: number | null;
}

export interface FareEstimateResponse {
  baseFare: number;
  surcharge: number;
  tollCharges: number;
  totalFare: number;
  trafficMultiplier: number;
  vehicleType: string;
  distanceKm: number;
  tollRoads: string[];
}

export interface SendOtpBody {
  phone: string;
}

export interface SendOtpResponse {
  success: boolean;
  message: string;
  /** Returned in demo mode; integrate SMS gateway for production */
  otp?: string;
}

export interface VerifyOtpBody {
  phone: string;
  otp: string;
}

export interface VerifyOtpResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}