# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bacchus Marsh Taxi Cab API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Submits a booking request and sends email and WhatsApp notification
 * @summary Submit a taxi booking
 */
export const createBookingBodyPassengersMax = 13;

export const CreateBookingBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  email: zod.string(),
  pickupAddress: zod.string(),
  dropoffAddress: zod.string(),
  vehicleType: zod.enum([
    "sedan",
    "suv",
    "silver_service",
    "six_seater",
    "maxi_taxi",
  ]),
  passengers: zod.number().min(1).max(createBookingBodyPassengersMax),
  pickupDate: zod.string(),
  pickupTime: zod.string(),
  isReturn: zod.boolean(),
  returnDate: zod.string().nullish(),
  returnTime: zod.string().nullish(),
  estimatedFare: zod.number().nullish(),
  notes: zod.string().nullish(),
  distanceKm: zod.number().nullish(),
  paymentMethod: zod.enum(["cash", "card", "cabcharge"]).nullish(),
});

export const CreateBookingResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  bookingId: zod.string(),
});

/**
 * Generates a 6-digit OTP and sends it to the provided phone number
 * @summary Send OTP to mobile number
 */
export const SendOtpBody = zod.object({
  phone: zod.string(),
});

export const SendOtpResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  otp: zod
    .string()
    .optional()
    .describe("Returned in demo mode; integrate SMS gateway for production"),
});

/**
 * Verifies the 6-digit OTP against the stored value
 * @summary Verify OTP for mobile number
 */
export const VerifyOtpBody = zod.object({
  phone: zod.string(),
  otp: zod.string(),
});

export const VerifyOtpResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * Calculates estimated fare based on distance and vehicle type
 * @summary Estimate taxi fare
 */
export const EstimateFareBody = zod.object({
  distanceKm: zod.number(),
  vehicleType: zod.enum([
    "sedan",
    "suv",
    "silver_service",
    "six_seater",
    "maxi_taxi",
  ]),
  pickupDate: zod.string().nullish(),
  pickupTime: zod.string().nullish(),
  tollRoads: zod
    .array(zod.string())
    .nullish()
    .describe(
      'Toll roads detected on the route (e.g. \"CityLink\", \"EastLink\", \"Westgate Tunnel\")',
    ),
  trafficRatio: zod
    .number()
    .nullish()
    .describe(
      "Optional live traffic multiplier from Google Maps (e.g. 1.25 means traffic is 25% slower than baseline)",
    ),
});

export const EstimateFareResponse = zod.object({
  baseFare: zod.number(),
  surcharge: zod.number(),
  tollCharges: zod.number(),
  totalFare: zod.number(),
  trafficMultiplier: zod.number(),
  vehicleType: zod.string(),
  distanceKm: zod.number(),
  tollRoads: zod.array(zod.string()),
});