# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bacchus Marsh Taxi Cab API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyVehicleType } from "./createBookingBodyVehicleType";

export interface CreateBookingBody {
  name: string;
  phone: string;
  email: string;
  pickupAddress: string;
  dropoffAddress: string;
  vehicleType: CreateBookingBodyVehicleType;
  /**
   * @minimum 1
   * @maximum 13
   */
  passengers: number;
  pickupDate: string;
  pickupTime: string;
  isReturn: boolean;
  returnDate?: string | null;
  returnTime?: string | null;
  estimatedFare?: number | null;
  notes?: string | null;
  distanceKm?: number | null;
  paymentMethod?: "cash" | "card" | "cabcharge" | null;
}